requirements.txt
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
numpy>=1.21.0
pillow>=8.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pandas>=1.3.0
tqdm>=4.62.0
tensorboard>=2.7.0